+++ src/index.ts
@@ -2,5 +2,5 @@ export * from './exec';
 export * from './execFile';
 export * from './fork';
 export * from './spawn';
-export * from './operators';
+export {trim} from './operators';
 export {spawnEnd, ShellError} from './util';