+++ test/fixtures/echoStderr.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+>&2 echo "ERR"