+++ test/fixtures/echostd.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+echo "HELLO"
+>&2 echo "WORLD"