+++ test/fixtures/forkError.js
@@ -1,0 +1,3 @@
+var execSync = require('child_process').execSync;
+
+execSync('mkdir test');