+++ test/fixtures/forkstd.js
@@ -1,0 +1,2 @@
+process.stdout.write('HELLO');
+process.stderr.write('WORLD');